% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.mvmeta.R
\name{bayes.mvmeta}
\alias{bayes.mvmeta}
\title{Fit Bayesian multivariate skew meta-regression models for individual patient data}
\usage{
bayes.mvmeta(Outcome, Trial, Patient, Treat, Covariate, tau,
  prior = list(), mcmc = list(), verbose = FALSE)
}
\arguments{
\item{Outcome}{aggregate mean of the responses for each arm of each study}

\item{Trial}{the identifier for a trial; it will internally be relabeled to contiguous/consecutive natural numbers starting from 1}

\item{Patient}{the patient number in each trial}

\item{Treat}{treatment applied; it will internally be relabeled to 0 and 1}

\item{Covariate}{aggregate covariates for the mean component}

\item{tau}{hyperparameter which determines the level of skewness}

\item{prior}{list of hyperparameters; when not given, algorithm will run in default setting}

\item{mcmc}{list of MCMC-related parameters: number of burn-ins (ndiscard), number of thinning(nskip), and posterior sample size (nkeep)}

\item{verbose}{logical variable for printing progress bar. Default to FALSE.}
}
\value{
a dataframe with input arguments, posterior samples, Metropolis algorithm acceptance rates, etc
}
\description{
This is a function for running the Markov chain Monte Carlo algorithm for the BFSHMVMR Model. The first six arguments are required.
}
\examples{
\dontrun{
data(df)
groupInfo <- list(c(0,1), c(2,3), c(4)) # define the variance structure
x <- df[,6:10]
fit <- bayes.mvmeta(df$y, df$sd, x, df$ids, df$iarm, df$npt, groupInfo,
			prior = list(c01=1.0e05, c02=4, nu=3),
			mcmc=list(ndiscard=2500,nskip=1,nkeep=10000))
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
