\name{BayesMeta-package}
\alias{BayesMeta-package}
\alias{BayesMeta}
\docType{package}
\title{
\packageTitle{BayesMeta}
}
\description{
\packageDescription{BayesMeta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesMeta}
\packageIndices{BayesMeta}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{BayesMeta}

Maintainer: \packageMaintainer{BayesMeta}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
